clean: [major] moved all verification passes to quantumion.compiler.analog.verify module and created new module for custom errors
