+++ quantumion/compiler/analog/error.py
@@ -1,0 +1,9 @@
+__all__ = [
+    "CanonicalFormError",
+]
+
+class CanonicalFormError(Exception):
+    """
+    Error class for canonical form
+    """
+    pass
